requirements.txt

runpod
torch
transformers
pillow
accelerate
einops
sentencepiece